--- documents/nbproject/genfiles.properties
@@ -1,8 +1,0 @@
-build.xml.data.CRC32=45685caf
-build.xml.script.CRC32=70fc3611
-build.xml.stylesheet.CRC32=651128d4@1.77.1.1
-# This file is used by a NetBeans-based IDE to track changes in generated files such as build-impl.xml.
-# Do not edit this file. You may delete it but then the IDE will never regenerate such files for you.
-nbproject/build-impl.xml.data.CRC32=45685caf
-nbproject/build-impl.xml.script.CRC32=a1660a6d
-nbproject/build-impl.xml.stylesheet.CRC32=99ea4b56@1.77.1.1
