--- documents/src/conf/MANIFEST.MF
@@ -1,2 +1,0 @@
-Manifest-Version: 1.0
-
