+++ project/nbproject/private/private.properties
@@ -1,0 +1,7 @@
+deploy.ant.properties.file=C:\\Users\\samue\\AppData\\Roaming\\NetBeans\\8.2rc\\config\\GlassFishEE6\\Properties\\gfv31589038235.properties
+j2ee.platform.is.jsr109=true
+j2ee.server.domain=C:/Users/samue/AppData/Roaming/NetBeans/8.2rc/config/GF_4.1.1/domain1
+j2ee.server.home=C:/Program Files/glassfish-4.1.1/glassfish
+j2ee.server.instance=[C:\\Program Files\\glassfish-4.1.1\\glassfish;C:\\Program Files\\glassfish-4.1.1\\glassfish\\domains\\domain1]deployer:gfv3ee6wc:localhost:4848
+j2ee.server.middleware=C:/Program Files/glassfish-4.1.1
+user.properties.file=C:\\Users\\samue\\AppData\\Roaming\\NetBeans\\8.2rc\\build.properties
