+++ project/src/conf/MANIFEST.MF
@@ -1,0 +1,2 @@
+Manifest-Version: 1.0
+
